+++ betfairlightweight/endpoints/racecard.py
@@ -1,5 +1,5 @@
 import re
-import datetime
+import time
 import requests
 from typing import Union, List
 
@@ -41,7 +41,7 @@ class RaceCard(BaseEndpoint):
         market_ids: list,
         data_entries: str = None,
         session: requests.Session = None,
-        lightweight: bool = True,
+        lightweight: bool = False,
     ) -> Union[list, List[resources.RaceCard]]:
         """
         Returns a list of race cards based on market ids provided.
@@ -59,11 +59,11 @@ class RaceCard(BaseEndpoint):
                 "APIClient.race_card.login()"
             )
         params = self.create_race_card_req(market_ids, data_entries)
-        (response, elapsed_time) = self.request(
+        (response, response_json, elapsed_time) = self.request(
             "raceCard", params=params, session=session
         )
         return self.process_response(
-            response, resources.RaceCard, elapsed_time, lightweight
+            response, response_json, resources.RaceCard, elapsed_time, lightweight
         )
 
     def get_race_result(
@@ -89,16 +89,18 @@ class RaceCard(BaseEndpoint):
                 "APIClient.race_card.login()"
             )
         params = self.create_race_result_req(market_ids, data_entries)
-        (response, elapsed_time) = self.request(
+        (response, response_json, elapsed_time) = self.request(
             "raceResults", params=params, session=session
         )
-        return self.process_response(response, None, elapsed_time, lightweight)
+        return self.process_response(
+            response, response_json, None, elapsed_time, lightweight
+        )
 
     def request(
         self, method: str = None, params: dict = None, session: requests.Session = None
     ) -> (dict, float):
         session = session or self.client.session
-        date_time_sent = datetime.datetime.utcnow()
+        time_sent = time.time()
         url = "%s%s" % (self.url, method)
         try:
             response = session.get(url, params=params, headers=self.headers)
@@ -106,15 +108,15 @@ class RaceCard(BaseEndpoint):
             raise APIError(None, method, params, e)
         except Exception as e:
             raise APIError(None, method, params, e)
-        elapsed_time = (datetime.datetime.utcnow() - date_time_sent).total_seconds()
+        elapsed_time = time.time() - time_sent
 
         check_status_code(response)
         try:
-            response_data = json_loads(response.text)
+            response_json = json_loads(response.text)
         except ValueError:
             raise InvalidResponse(response.text)
 
-        return response_data, elapsed_time
+        return response, response_json, elapsed_time
 
     @staticmethod
     def create_race_card_req(market_ids: list, data_entries: str) -> dict: