+++ tests/unit/test_baseendpoint.py
@@ -49,8 +49,8 @@ class BaseEndPointTest(unittest.TestCase):
             headers=mock_request_headers,
             timeout=(3.05, 16),
         )
-        assert response[0] == mock_response.json()
-        assert isinstance(response[1], float)
+        assert response[1] == mock_response.json()
+        assert isinstance(response[2], float)
 
     @mock.patch("betfairlightweight.endpoints.baseendpoint.BaseEndpoint.create_req")
     @mock.patch("betfairlightweight.baseclient.BaseClient.cert")
@@ -98,32 +98,33 @@ class BaseEndPointTest(unittest.TestCase):
             self.base_endpoint._error_handler(mock_response.json())
 
     def test_base_endpoint_process_response(self):
+        mock_response = mock.Mock()
         mock_resource = mock.Mock()
 
         response_list = [{}, {}]
         response = self.base_endpoint.process_response(
-            response_list, mock_resource, None, False
+            mock_response, response_list, mock_resource, None, False
         )
         assert type(response) == list
         assert response[0] == mock_resource()
 
         response_result_list = {"result": [{}, {}]}
         response = self.base_endpoint.process_response(
-            response_result_list, mock_resource, None, False
+            mock_response, response_result_list, mock_resource, None, False
         )
         assert type(response) == list
         assert response[0] == mock_resource()
 
         response_result = {"result": {}}
         response = self.base_endpoint.process_response(
-            response_result, mock_resource, None, False
+            mock_response, response_result, mock_resource, None, False
         )
         assert response == mock_resource()
 
         # lightweight tests
         response_list = [{}, {}]
         response = self.base_endpoint.process_response(
-            response_list, mock_resource, None, True
+            mock_response, response_list, mock_resource, None, True
         )
         assert response == response_list
 
@@ -131,22 +132,23 @@ class BaseEndPointTest(unittest.TestCase):
         base_endpoint = BaseEndpoint(client)
         response_list = [{}, {}]
         response = base_endpoint.process_response(
-            response_list, mock_resource, None, False
+            mock_response, response_list, mock_resource, None, False
         )
         assert type(response) == list
         assert response[0] == mock_resource()
 
     def test_base_endpoint_process_response_no_error(self):
         class MockResource:
-            def __init__(self, elapsed_time, hello):
+            def __init__(self, elapsed_time, hello, **kwargs):
                 self.elapsed_time = elapsed_time
                 self.hello = hello
 
+        mock_response = mock.Mock()
         mock_resource = MockResource
         response_list = [{"hello": "world"}]
 
         response = self.base_endpoint.process_response(
-            response_list, mock_resource, 0, False
+            mock_response, response_list, mock_resource, 0, False
         )
         assert type(response) == list
         assert (
@@ -160,11 +162,14 @@ class BaseEndPointTest(unittest.TestCase):
                 self.elapsed_time = elapsed_time
                 self.hello = hello
 
+        mock_response = mock.Mock()
         mock_resource = MockResource
         response_list = [{}]  # missing 'hello'
 
         with self.assertRaises(InvalidResponse):
-            self.base_endpoint.process_response(response_list, mock_resource, 0, False)
+            self.base_endpoint.process_response(
+                mock_response, response_list, mock_resource, 0, False
+            )
 
     def test_base_endpoint_url(self):
         assert self.base_endpoint.url == "%s%s" % (